+++ Plugin/Uplink/Source/UplinkEditor/Private/UplinkTools.h
@@ -31,6 +31,7 @@ namespace UplinkTools
 	void RegisterData(FUplinkToolRegistry& Registry);
 	void RegisterTests(FUplinkToolRegistry& Registry);
 	void RegisterStruct(FUplinkToolRegistry& Registry);
+	void RegisterEnum(FUplinkToolRegistry& Registry);
 	void RegisterPCG(FUplinkToolRegistry& Registry);
 	void RegisterGASP(FUplinkToolRegistry& Registry);
 	void RegisterAutoplay(FUplinkToolRegistry& Registry);